Add A,B,X,Y buttons and thumbstick inputs
refactor collection of inputs into an `Input` struct
