+++ hotham/src/schedule_functions/apply_haptic_feedback.rs
@@ -16,6 +16,8 @@ static HAPTIC_DURATION: i64 = 1e+8 as _; // 100ms
 /// }
 /// ```
 pub fn apply_haptic_feedback(xr_context: &mut XrContext, haptic_context: &mut HapticContext) {
+    let input = &xr_context.input;
+
     let haptic_duration = Duration::from_nanos(HAPTIC_DURATION);
     if haptic_context.left_hand_amplitude_this_frame != 0. {
         let event = HapticVibration::new()
@@ -23,13 +25,9 @@ pub fn apply_haptic_feedback(xr_context: &mut XrContext, haptic_context: &mut Ha
             .frequency(HAPTIC_FREQUENCY)
             .duration(haptic_duration);
 
-        xr_context
+        input
             .haptic_feedback_action
-            .apply_feedback(
-                &xr_context.session,
-                xr_context.left_hand_subaction_path,
-                &event,
-            )
+            .apply_feedback(&xr_context.session, input.left_hand_subaction_path, &event)
             .expect("Unable to apply haptic feedback!");
 
         // Reset the value
@@ -42,13 +40,9 @@ pub fn apply_haptic_feedback(xr_context: &mut XrContext, haptic_context: &mut Ha
             .frequency(HAPTIC_FREQUENCY)
             .duration(haptic_duration);
 
-        xr_context
+        input
             .haptic_feedback_action
-            .apply_feedback(
-                &xr_context.session,
-                xr_context.right_hand_subaction_path,
-                &event,
-            )
+            .apply_feedback(&xr_context.session, input.right_hand_subaction_path, &event)
             .expect("Unable to apply haptic feedback!");
 
         // Reset the value
